feat: added onRemove callback.
The onRemove callback will be called whenever the user manually closes
the flash message or when it is removed due to the duration expiring.
The callback will receive the flash message which is removed and the
reason for removal.

Added more comments throughout out the application.

Tested the behavior of the onClick from the default actions, and as
as small refactoring the default actions now create the flash message
by spreading the config `...config` instead of doing each prop manually.

Closes: #10